+++ spec/lib/oauth/client_credentials/creator_spec.rb
@@ -5,6 +5,10 @@ class Doorkeeper::OAuth::ClientCredentialsRequest
     let(:client) { FactoryBot.create :application }
     let(:scopes) { Doorkeeper::OAuth::Scopes.from_string('public') }
 
+    before do
+      default_scopes_exist :public
+    end
+
     it 'creates a new token' do
       expect do
         subject.call(client, scopes)